Enhanced the no mask process in the pipeline added 255 to create a white mask and added dtype

+++ src/library/image_manipulation/mask_manager.py
@@ -185,7 +185,8 @@ class MaskManager:
                 del mask
             else:
                 img = np.array(img)
-                merged_img = np.zeros_like(img.astype(np.uint8))
+                merged_img = np.zeros_like(img)
+                merged_img = merged_img.astype(np.uint8)
                 merged_img = 255
             cv2.imwrite(maskpath, merged_img)
